Extract structs and classes to files
